+++ CountlyConsentManager.h
@@ -20,6 +20,7 @@
 @property (nonatomic, readonly) BOOL consentForAttribution;
 @property (nonatomic, readonly) BOOL consentForStarRating;
 @property (nonatomic, readonly) BOOL consentForAppleWatch;
+@property (nonatomic, readonly) BOOL consentForPerformanceMonitoring;
 
 + (instancetype)sharedInstance;
 - (void)giveConsentForFeatures:(NSArray *)features;