+++ CountlyPersistency.m
@@ -315,7 +315,14 @@ NSString* const kCountlyCustomCrashLogFileName = @"CountlyCustomCrash.log";
         saveData = [NSKeyedArchiver archivedDataWithRootObject:@{kCountlyQueuedRequestsPersistencyKey: self.queuedRequests}];
     }
 
-    [saveData writeToFile:[self storageFileURL].path atomically:YES];
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wunused-variable"
+
+    BOOL writeResult = [saveData writeToFile:[self storageFileURL].path atomically:YES];
+    COUNTLY_LOG(@"Result of writing data to file: %d", writeResult);
+
+#pragma clang diagnostic pop
+
     [CountlyCommon.sharedInstance finishBackgroundTask];
 }